Added Clock library to Wifi_Clock library. Renamed 'clock.cpp' to 'timing.cpp'. Started unifying time format.
