--- arduino/Clock/README.md
@@ -1,15 +1,0 @@
-# Clock
-The Clock library is an Arduino library written originally to supplement the Wifi Clock kit developed by MicroNote. This library can be used on any esp8266 board for any purpose since it does not rely on external hardware.
-
-## Documentation
-Documentation for the Clock library is available on our website [here](https://micronote.tech/wificlock_product/index.html), or on our [github page](https://github.com/MicroNotebook/wifi_clock/tree/master/arduino).
-
-## Download
-The lastest source code version of the Library is always available from the 
-[WifiClock page](https://github.com/MicroNotebook/wifi_clock/tree/master/arduino).
-
-## Installation
-The library can be installed using the [standard Arduino library install procedure](http://arduino.cc/en/Guide/Libraries#.UwxndHX5PtY).
-
-## Dependencies
-This library needs the *Ticker* libraries installed to function correctly. These libraries can be installed with the **Library Manager** in Arduino (see [standard Arduino library install procedure](http://arduino.cc/en/Guide/Libraries#.UwxndHX5PtY)).
