--- arduino/Clock/keywords.txt
@@ -1,23 +1,0 @@
-#######################################
-# Syntax Coloring Map For WifiClock
-#######################################
-
-#######################################
-# Datatypes (KEYWORD1)
-#######################################
-
-Clock	KEYWORD1
-Time	KEYWORD1
-
-#######################################
-# Methods and Functions (KEYWORD2)
-#######################################
-
-getTime	KEYWORD2
-setTime	KEYWORD2
-startClock	KEYWORD2
-stopClock	KEYWORD2
-
-#######################################
-# Constants (LITERAL1)
-#######################################
