--- arduino/Clock/library.properties
@@ -1,11 +1,0 @@
-name=Clock
-version=1.0.0
-author=Holden Rahn <holden.rahn@gmail.com>
-maintainer=Holden Rahn <holden.rahn@gmail.com>
-sentence=A library to control and maintain time on the esp8266.
-paragraph=This library allows for easy and accurate updating of time. This library was originally designed to be used with the WifiClock library.
-category=Timing
-url=https://micronote.tech/wificlock_product/index.html
-architectures=esp8266
-includes=Clock.h
-dot_a_linkage=true
