+++ arduino/WifiClock/src/clock.cpp
@@ -1,100 +1,179 @@
-#include "WifiClock.h"
+#include "Clock.h"
+#include "Arduino.h"
 
-void WifiClock::start_clock(void)
+Clock::Clock(byte sec, byte min, byte hr, byte day, byte month, int year)
 {
-	_clock.startClock();
+	this->setTime(sec, min, hr, day, month, year);
 }
 
-void WifiClock::stop_clock(void)
+Clock::Clock(Time time)
 {
-	_clock.stopClock();
+	this->setTime(time);
 }
 
-Time WifiClock::get_time(bool military)
+Clock::Clock(void)
 {
-	return _clock.getTime(military);
+	/* NOP */
 }
 
-void WifiClock::set_time(Time time)
+Time Clock::getTime(void)
 {
-	_clock.setTime(time);
+	return _t;
 }
 
-void WifiClock::set_time(byte sec, byte min, byte hr, byte day, byte month, int year)
+void Clock::setTime(Time time)
 {
-	_clock.setTime(sec, min, hr, day, month, year);
-}
-
-void WifiClock::display_time(bool military, bool secs, bool right)
-{
-	// Configure Time
-	Time t = _clock.getTime(military);
-	byte hour = t.hours;
-	byte min = t.minutes;
-	byte sec = t.seconds;
-	
-	// Display Time
-	if (secs) {			// seconds
-		_multi_display(hour / 10, hour % 10, min / 10, min % 10, sec / 10, sec % 10);
+	if (time.seconds < 60 && time.seconds >= 0) {
+		_t.seconds = time.seconds;
+	}
+	if (time.minutes < 60 && time.minutes >= 0) {
+		_t.minutes = time.minutes;
+	}
+	if (time.hours < 24 && time.hours >= 0) {
+		_t.hours = time.hours;
+	}
+	if (time.days > 0 && time.days <= 31) {
+		_t.days = time.days;
+	}
+	if (time.months > 0 && time.months <= 12) {
+		_t.months = time.months;
+	}
+	if (time.years > -1000 && time.years <= 9999) {
+		_t.years = time.years;
+	}
+	if (_t.hours >= 12) {
+		_t.PM = true;
 	} else {
-		if (right) {	// no seconds, right aligned
-			_multi_display(255, 255, hour / 10, hour % 10, min / 10, min % 10);
-			
-		} else {		// no seconds, left aligned
-			_multi_display(hour / 10, hour % 10, min / 10, min % 10, 255, 255);
-		}
-		_curr_type = 0;
+		_t.PM = false;
 	}
 }
 
-void WifiClock::display_date(byte pos)
+void Clock::setTime(byte sec, byte min, byte hr, byte day, byte month, int year)
 {
-  // Configure Date
-  Time t = _clock.getTime(false);
-  byte day = t.days;
-  byte month = t.months;
-  
-  // Display Date
-  if (pos <= 2) {
-    _lc.setDigit(0, pos++, day % 10, false);
-    _lc.setDigit(0, pos++, day / 10, false);
-    _lc.setDigit(0, pos++, month % 10, false);
-    _lc.setDigit(0, pos, month / 10, false);
-	_curr_type = 0;
-  }
+	if (sec < 60 && sec >= 0) {
+		_t.seconds = sec;
+	}
+	if (min < 60 && min >= 0) {
+		_t.minutes = min;
+	}
+	if (hr < 24 && hr >= 0) {
+		_t.hours = hr;
+	}
+	if (day > 0 && day <= 31) {
+		_t.days = day;
+	}
+	if (month > 0 && month <= 12) {
+		_t.months = month;
+	}
+	if (year > -1000 && year <= 9999) {
+		_t.years = year;
+	}
+	if (_t.hours >= 12) {
+		_t.PM = true;
+	} else {
+		_t.PM = false;
+	}
 }
 
-void WifiClock::display_day(byte pos)
+void Clock::startClock(void)
 {
-  byte day = (_clock.getTime(false)).days;
-  if (pos <= 4) {
-    _lc.setDigit(0, pos++, day % 10, false);
-    _lc.setDigit(0, pos, day / 10, false);
-	_curr_type = 0; 
-  }
+	_timer.attach(1.0, [this](void){ this->_increase_second(); });
 }
 
-void WifiClock::display_month(byte pos)
+void Clock::stopClock(void)
 {
-  byte month = (_clock.getTime(false)).months;
-  if (pos <= 4) {
-    _lc.setDigit(0, pos++, month % 10, false);
-    _lc.setDigit(0, pos, month / 10, false);
-	_curr_type = 0;
-  }
+	_timer.detach();
 }
 
-void WifiClock::display_year(byte pos)
+void Clock::_increase_year(void)
 {
-  // Configure Date
-  int year = (_clock.getTime(false)).years;
-  
-  // Display Date
-  if (pos <= 2) {
-    for(int i = pos ; i < pos + 4; i++) {
-      _lc.setDigit(0, i, year % 10, false);
-      year /= 10;
-    }
-	_curr_type = 0;
-  }
+	_t.years++;
+	if (_t.years > 9999) {
+		_t.years = 9999;
+	}
+}
+
+void Clock::_increase_month(void)
+{
+	_t.months++;
+	if (_t.months > 12) {
+		_t.months = 1;
+		_increase_year();
+	}
+}
+
+void Clock::_increase_day(void)
+{
+	_t.days++;
+	switch (_t.months)
+	{ 
+		// 31 day months
+		case 1:
+		case 3:
+		case 5:
+		case 7:
+		case 8:
+		case 10:
+		case 12:
+			if (_t.days > 31) {
+				_t.days = 1;
+				_increase_month();
+			}
+			break;
+
+		// 30 day months
+		case 4:
+		case 6:
+		case 9:
+		case 11:
+			if (_t.days > 30) {
+				_t.days = 1;
+				_increase_month();
+			}
+			break;
+
+		// February
+		default:
+			if ((_t.years % 4 == 0 && _t.years % 100 != 0) || (_t.years % 400 == 0)) {  // if leap year
+				if (_t.days > 29) {
+				  _t.days = 1;
+				  _increase_month();
+				}
+			} else {
+				if (_t.days > 28) {
+				  _t.days = 1;
+				  _increase_month();
+				}
+			}
+	}
+}
+
+void Clock::_increase_hour(void)
+{
+	_t.hours++;
+	if (_t.hours > 24) {
+		_t.hours = 0;
+		_increase_day();
+	}
+	if (_t.hours >= 12) {
+		_t.PM = true;
+	}
+}
+
+void Clock::_increase_minute(void)
+{
+	_t.minutes++;
+	if (_t.minutes >= 60) {
+		_t.minutes = 0;
+		_increase_hour();
+	}
+}
+
+void Clock::_increase_second(void)
+{
+	_t.seconds++;
+	if (_t.seconds >= 60) {
+		_t.seconds = 0;
+		_increase_minute();
+	}
 }
